requirements.txt
python-dotenv
pandas
numpy
matplotlib
seaborn
scikit-learn
pymongo
certifi
pymongo[srv]
dill
pyaml
#-e .